集成es

+++ xxl-job-admin/src/main/java/com/xxl/job/admin/controller/JobLogController.java
@@ -120,7 +120,7 @@ public class JobLogController {
     }
 
     @RequestMapping("/logDetailPage")
-    public String logDetailPage(int id, Model model) {
+    public String logDetailPage(long id, Model model) {
 
         // base check
         ReturnT<String> logStatue = ReturnT.SUCCESS;
